backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
pillow==10.2.0
tensorflow==2.16.1
numpy==1.26.4
joblib==1.3.2
requests==2.31.0
google-genai==0.3.0
redis==5.0.3
celery==5.3.6
